# language-java/jws/src/main/java/fr/epita/assistant/jws/presentation/rest/request/JoinGameRequest.java
package fr.epita.assistant.jws.presentation.rest.request;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@NoArgsConstructor @Getter @Setter @AllArgsConstructor
public class JoinGameRequest {
    public String name;
}